refactor CategoriesList tests (#1872)
* refactor CategoriesList tests

* remove test case

* add level option

* move tests to tests/unit
