+++ src/tests/unit/pages/home/categories-list/categories-list.variables.js
@@ -1,0 +1,8 @@
+const mockCategory = [
+  { lang: 'ua', value: 'test' },
+  { lang: 'en', value: 'test' }
+];
+
+const etalonString = 'catalog/test';
+
+export { mockCategory, etalonString };
